Updated ChannelHeader and MessageWrapper to match the changes to TextFormatting
