+++ webapp/components/message_wrapper.jsx
@@ -2,6 +2,7 @@
 // See License.txt for license information.
 
 import * as TextFormatting from 'utils/text_formatting.jsx';
+import * as Utils from 'utils/utils.jsx';
 
 import React from 'react';
 
@@ -10,9 +11,19 @@ export default class MessageWrapper extends React.Component {
         super(props);
         this.state = {};
     }
+
     render() {
         if (this.props.message) {
-            return <div dangerouslySetInnerHTML={{__html: TextFormatting.formatText(this.props.message, this.props.options)}}/>;
+            const options = Object.assign({}, this.props.options, {
+                siteURL: Utils.getSiteURL()
+            });
+
+            return (
+                <div
+                    onClick={Utils.handleFormattedTextClick}
+                    dangerouslySetInnerHTML={{__html: TextFormatting.formatText(this.props.message, options)}}
+                />
+            );
         }
 
         return <div/>;
