+++ webapp/components/post_view/components/post_message_view.jsx
@@ -3,7 +3,6 @@
 
 import React from 'react';
 
-import {browserHistory} from 'react-router/es6';
 import * as TextFormatting from 'utils/text_formatting.jsx';
 import * as Utils from 'utils/utils.jsx';
 
@@ -31,31 +30,6 @@ export default class PostMessageView extends React.Component {
         return false;
     }
 
-    handleClick(e) {
-        // TODO should this be here or somewhere else? it's already copied from TextFormatting
-        const mentionAttribute = e.target.getAttributeNode('data-mention');
-        const hashtagAttribute = e.target.getAttributeNode('data-hashtag');
-        const linkAttribute = e.target.getAttributeNode('data-link');
-
-        if (mentionAttribute) {
-            e.preventDefault();
-
-            Utils.searchForTerm(mentionAttribute.value);
-        } else if (hashtagAttribute) {
-            e.preventDefault();
-
-            Utils.searchForTerm(hashtagAttribute.value);
-        } else if (linkAttribute) {
-            const MIDDLE_MOUSE_BUTTON = 1;
-
-            if (!(e.button === MIDDLE_MOUSE_BUTTON || e.altKey || e.ctrlKey || e.metaKey || e.shiftKey)) {
-                e.preventDefault();
-
-                browserHistory.push(linkAttribute.value);
-            }
-        }
-    }
-
     render() {
         if (!this.props.enableFormatting) {
             return <span>{this.props.message}</span>;
@@ -63,12 +37,12 @@ export default class PostMessageView extends React.Component {
 
         const options = {
             emojis: this.props.emojis,
-            siteURL: global.mm_config.SiteURL || window.location.origin
+            siteURL: Utils.getSiteURL()
         };
 
         return (
             <span
-                onClick={this.handleClick}
+                onClick={Utils.handleFormattedTextClick}
                 dangerouslySetInnerHTML={{__html: TextFormatting.formatText(this.props.message, options)}}
             />
         );
